temp save

+++ NextWeek/vec3.h
@@ -51,7 +51,7 @@ inline std::ostream& operator<<(std::ostream &os, const vec3 &t) {
 }
 
 __host__ __device__ inline void vec3::make_unit_vector() {
-    float k = 1.0 / sqrt(e[0]*e[0] + e[1]*e[1] + e[2]*e[2]);
+    float k = 1.0f / sqrtf(e[0]*e[0] + e[1]*e[1] + e[2]*e[2]);
     e[0] *= k; e[1] *= k; e[2] *= k;
 }
 
@@ -130,7 +130,7 @@ __host__ __device__ inline vec3& vec3::operator*=(const float t) {
 }
 
 __host__ __device__ inline vec3& vec3::operator/=(const float t) {
-    float k = 1.0/t;
+    float k = 1.0f / t;
 
     e[0]  *= k;
     e[1]  *= k;
